Add invalid callback function.
Add an optional callback function to handle invalid requests.
This is for DoS protection so that the application can be notified of
invalid requests and handle them itself.

The automatic response code is now determined in request_receiver and
sent from http_server.
The http_connection receive function now only receives and no longer
sends responses.